emscripten ivec2 fix

+++ addons/ofxEmscripten/src/ofxAppEmscriptenWindow.h
@@ -36,9 +36,9 @@ public:
 	// void	setWindowPosition(int x, int y);
 	void	setWindowShape(int w, int h);
 
-	glm::vec2	getWindowPosition();
-	glm::vec2	getWindowSize();
-	glm::vec2	getScreenSize();
+	glm::ivec2 getWindowPosition();
+	glm::ivec2 getWindowSize();
+	glm::ivec2 getScreenSize();
 
 	
 	ofOrientation	getOrientation();
